requirements.txt
boto3==1.20.26
picamera==1.13
python-dotenv==0.19.2